[deckhelper] fix minor bugs
This commit fixes two bugs in deckhelper:

First, the number of cards by rarity was being miscounted. It was
adding one per card without taking into account how many copies
of that card were needed.

The second bug is that the parser was failing to parse some mtg:a
decks, specifically, the ones that had Command, Deck, Sideboard or
empty strings.