Update MTG Parsing format.
In the latest MTGA Update there were some changes in the way the logs
and the resource files were stored:

The JSON format for the GetPlayerCardsV3 message changed. Previously it
was a JSON with just `cardid` and its count: `{ "1": 3, "101": 4 }`.
Now that message is wrapped in another json, with an `id` and a `payload`:
`{ "id": 13, "payload": { "1": 3, "101": 4 } }`. Also, previously the
message used to be in its own line, but now it can be in the same line
as the text saying the message.

The code to fix that last issue is a little bit hacky: it will check if
each line starts with the message name, and if so, will use a combined
io.Reader to read from the start of the JSON (if it starts in that line)
and the rest of the original reader.

The last breaking change was in the resources file. Aparently, MTGA
deprecated the use of "langkey" to identify different languages in the
texts resource file. All the previous values are now `null`. It looks
like they moved to use the iso code of the language to be more specific.

Lastly, there's a minor change that makes the code use the last-parsed
collection instead of the first one. A single log file can contain a
collection progression of multiple hours for some people.