[deckhelper] Add set selection to allow historic
This commit modifies the logic in deckhelper allowing the user to
specify the list of enabled sets that can be used to calculate
crafting, allowing the user to specify decks for historic.

Users can invoke this command with the flag "--sets" and a list of
comma-separated names of expansions (encoded in the three letter
format used by Magic). There are two special-cases: STD and ALL.
The former will add all the sets in the current standard rotation
(harcoded...) and the latter will enable all sets (hardcoded...)

To enable this behavior, the code had to be refactored: a new struct
called deckhelper will hold the user's collection, the card library and
the map of enabled sets.