[limitedcollection] Add booster tracking program.
This commit adds a new program: limitedcollection, which can be used to
read all the boosters that you opened during the last time you used MTG
Arena.

This commit adds the functionallity to search for specific messages
inside the MTG Arena logs in the collection_finder library. An example
of this is that now collectionexporter can display the user's inventory.